requirements.txt
asgiref==3.8.1
Django==5.1.5
pillow==11.1.0
python-decouple==3.8
sqlparse==0.5.3
whitenoise==6.8.2
